Update functions.py
added device_type_ext. Device type extension is used to append the measured port to the saved file. For example. If you take an IV curve of nTron A1 was the IV at the gate or at the drain? Add 'port' to the 'Save File' dictionary to make that distinction.