update_table, log_data_to_database, Data update
update_table() can easily run sql commands in the format "UPDATE *table* SET *set_col* WHERE *conditional*".
Data class now has optional arguments `table_name` and `connection` in __init__ for adding a connection to an existing table in a database. If a connection is properly set up, store() will save data to a database, given that the column names in the table are the same as variable names in Data class.
By default, a SYSTEM variable sets the increment for committing to the database, but if you remove this system variable then the commit increment will be set to `save_increment`